Remove audio compatibility services and restore canonical sound owners

+++ src/JSystem/JAudio2/JAISoundHandles.cpp
@@ -2,7 +2,7 @@
 #include "JSystem/JAudio2/JAISeqDataMgr.hpp"
 #include "JSystem/JAudio2/JAISound.hpp"
 
-JAISoundHandle* JAISoundHandles::getHandleSoundID(JAISoundID id) {
+JAISoundHandle *JAISoundHandles::getHandleSoundID(JAISoundID id) {
     for (int i = 0; i < mNumHandles; i++) {
         if (mHandles[i].isSoundAttached()) {
             if (mHandles[i]->mSoundID == id) {
@@ -14,7 +14,7 @@ JAISoundHandle* JAISoundHandles::getHandleSoundID(JAISoundID id) {
     return nullptr;
 }
 
-JAISoundHandle* JAISoundHandles::getFreeHandle() {
+JAISoundHandle *JAISoundHandles::getFreeHandle() {
     for (int i = 0; i < mNumHandles; i++) {
         if (!mHandles[i].isSoundAttached()) {
             return &mHandles[i];
@@ -24,10 +24,10 @@ JAISoundHandle* JAISoundHandles::getFreeHandle() {
     return nullptr;
 }
 
-JAISoundHandle* JAISoundHandles::getHandleUserData(u32 addr) {
+JAISoundHandle *JAISoundHandles::getHandleUserData(uintptr_t addr) {
     for (int i = 0; i < mNumHandles; i++) {
         if (mHandles[i].isSoundAttached()) {
-            if ((u32)mHandles[i]->getUserData() == addr) {
+            if (mHandles[i]->getUserData() == addr) {
                 return &mHandles[i];
             }
         }
@@ -36,7 +36,7 @@ JAISoundHandle* JAISoundHandles::getHandleUserData(u32 addr) {
     return nullptr;
 }
 
-void JAISoundHandles::setPos(const TVec3f& pos) {
+void JAISoundHandles::setPos(const TVec3f &pos) {
     for (int i = 0; i < mNumHandles; i++) {
         if (mHandles[i].isSoundAttached()) {
             mHandles[i].getSound()->setPos(pos);
